# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Сервис помогает искать специалистов различных профессий...",
        "title": "Unclaimeds API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8585",
    "basePath": "/api",
    "paths": {
        "/account/sessions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает список всех активных сессий текущего пользователя.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Сессии"
                ],
                "summary": "Получить активные сессии",
                "responses": {
                    "200": {
                        "description": "Список сессий",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/auth/login/send-code": {
            "post": {
                "description": "Отправляет 6-значный код на email для входа в систему без пароля.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Аутентификация"
                ],
                "summary": "Отправить код для входа",
                "parameters": [
                    {
                        "description": "Email пользователя",
                        "name": "email",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.LoginEmailRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Код отправлен",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/auth/login/verify-code": {
            "post": {
                "description": "Проверяет 6-значный код, присланный на email, и авторизует пользователя.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Аутентификация"
                ],
                "summary": "Проверить код и войти",
                "parameters": [
                    {
                        "description": "Email и код",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.VerifyEmailCodeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Вход выполнен успешно",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/auth/password": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Позволяет авторизованному пользователю изменить свой пароль.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Пароли"
                ],
                "summary": "Изменить пароль",
                "responses": {
                    "200": {
                        "description": "Пароль успешно изменен",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/auth/password/reset": {
            "post": {
                "description": "Отправляет на указанный email ссылку для сброса пароля.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Пароли"
                ],
                "summary": "Запрос на сброс пароля",
                "parameters": [
                    {
                        "description": "Email пользователя для сброса пароля",
                        "name": "email",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.LoginEmailRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ссылка успешно отправлена",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/auth/password/reset/confirm": {
            "post": {
                "description": "Устанавливает новый пароль, используя токен сброса.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Пароли"
                ],
                "summary": "Сброс пароля",
                "responses": {
                    "200": {
                        "description": "Пароль успешно обновлён",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/auth/password/reset/page": {
            "get": {
                "description": "Возвращает информацию для страницы сброса пароля, используя токен.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Пароли"
                ],
                "summary": "Получить страницу сброса пароля",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Токен сброса пароля",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о сбросе",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/auth/resend-code": {
            "post": {
                "description": "Отправляет новый 6-значный код на email для верификации.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Верификация"
                ],
                "summary": "Повторная отправка кода",
                "parameters": [
                    {
                        "description": "Email пользователя",
                        "name": "email",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.ResendCodeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Новый код отправлен",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/auth/session": {
            "get": {
                "description": "Проверяет наличие и валидность сессии пользователя.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Сессии"
                ],
                "summary": "Проверка сессии",
                "responses": {
                    "200": {
                        "description": "Сессия активна",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/auth/sessions/revoke": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Аннулирует указанную сессию пользователя по ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Сессии"
                ],
                "summary": "Отменить сессию",
                "responses": {
                    "200": {
                        "description": "Сессия успешно отменена",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/auth/signout": {
            "post": {
                "description": "Деавторизует текущего пользователя, удаляя сессию.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Аутентификация"
                ],
                "summary": "Выход из системы",
                "responses": {
                    "200": {
                        "description": "Успешный выход",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/auth/signup": {
            "post": {
                "description": "Создает нового пользователя с логином и паролем.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Регистрация"
                ],
                "summary": "Регистрация нового пользователя",
                "parameters": [
                    {
                        "description": "Данные для регистрации",
                        "name": "signupRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.SignUpRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешная регистрация",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/auth/signup/send-code": {
            "post": {
                "description": "Отправляет 6-значный код на email для упрощенной регистрации.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Регистрация"
                ],
                "summary": "Отправить код для регистрации",
                "parameters": [
                    {
                        "description": "Email нового пользователя",
                        "name": "email",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.SignupEmailRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Код отправлен",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/auth/signup/verify-code": {
            "post": {
                "description": "Проверяет 6-значный код и автоматически создает аккаунт.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Регистрация"
                ],
                "summary": "Проверить код и зарегистрировать пользователя",
                "parameters": [
                    {
                        "description": "Email и код",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.VerifyEmailCodeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Регистрация прошла успешно",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "description": "Проводит авторизацию пользователя по логину и паролю.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Аутентификация"
                ],
                "summary": "Авторизация пользователя",
                "parameters": [
                    {
                        "description": "Данные для входа",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный вход",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.LoginEmailRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "domain.LoginRequest": {
            "type": "object",
            "properties": {
                "password_hash": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "domain.ResendCodeRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "domain.SignUpRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "description": "Электронная почта пользователя",
                    "type": "string"
                },
                "first_name": {
                    "description": "Имя пользователя",
                    "type": "string"
                },
                "last_name": {
                    "description": "Фамилия пользователя",
                    "type": "string"
                },
                "password_hash": {
                    "description": "Хэш нового пароля",
                    "type": "string"
                },
                "username": {
                    "description": "Имя пользователя для входа",
                    "type": "string"
                }
            }
        },
        "domain.SignupEmailRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "domain.VerifyEmailCodeRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                }
            }
        },
        "utils.Response": {
            "type": "object",
            "properties": {
                "body": {
                    "description": "Тело ответа (может содержать любые данные или ErrorResponse)"
                },
                "statusCode": {
                    "description": "Код статуса HTTP-ответа",
                    "type": "integer"
                }
            }
        }
    }
}